+++ src/Domain/User/Repository/UserRepositoryInterface.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace App\Domain\User\Repository;
 
 use App\Domain\Common\Specification\SpecificationInterface;
-use App\Domain\User\Entity\UserViewInterface;
+use Broadway\ReadModel\SerializableReadModel;
 
 interface UserRepositoryInterface
 {
-    public function getOneOrNull(SpecificationInterface $specification): ?UserViewInterface;
+    public function getOneOrNull(SpecificationInterface $specification): ?SerializableReadModel;
 }